move rifaValidation to pages dir
